refactor(codegen): prepare for v2
