+++ codegen/src/v1/aws_conv.rs
@@ -1,7 +1,7 @@
-use crate::dto::RustTypes;
-use crate::ops::is_op_input;
-use crate::ops::Operations;
-use crate::rust;
+use super::dto::RustTypes;
+use super::ops::is_op_input;
+use super::ops::Operations;
+use super::rust;
 
 use std::format as f;
 use std::ops::Not;
