+++ codegen/src/v1/aws_proxy.rs
@@ -1,6 +1,6 @@
-use crate::dto::RustTypes;
-use crate::ops::Operations;
-use crate::rust;
+use super::dto::RustTypes;
+use super::ops::Operations;
+use super::rust;
 
 use std::format as f;
