+++ codegen/src/v1/xml.rs
@@ -1,7 +1,7 @@
-use crate::dto::RustTypes;
-use crate::ops::Operations;
-use crate::rust;
-use crate::rust::default_value_literal;
+use super::dto::RustTypes;
+use super::ops::Operations;
+use super::rust;
+use super::rust::default_value_literal;
 
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
 use std::ops::Not;
